Refactor world-utils module: consolidate exports and remove unused imports; update world-update to reflect new import structure
