+++ games/tribe/src/game/ai/behavior-tree/behaviors/desperate-attack-behavior.ts
@@ -1,6 +1,6 @@
 import { HumanEntity } from '../../../entities/characters/human/human-types';
 import { UpdateContext } from '../../../world-types';
-import { findWeakCannibalismTarget, findClosestEntity } from '../../../utils/world-utils';
+import { findWeakCannibalismTarget, findClosestEntity } from '../../../utils';
 import { AI_DESPERATE_ATTACK_HUNGER_THRESHOLD, HUMAN_INTERACTION_RANGE } from '../../../world-consts';
 import { BehaviorNode, NodeStatus } from '../behavior-tree-types';
 import { Sequence, ConditionNode, ActionNode } from '../nodes';
