+++ games/tribe/src/game/ai/behavior-tree/behaviors/jealousy-attack-behavior.ts
@@ -1,6 +1,6 @@
 import { HumanEntity } from '../../../entities/characters/human/human-types';
 import { UpdateContext } from '../../../world-types';
-import { findPartnerProcreatingWithStranger } from '../../../utils/world-utils';
+import { findPartnerProcreatingWithStranger } from '../../../utils';
 import { AI_JEALOUSY_PROCREATION_TRIGGER_RADIUS } from '../../../world-consts';
 import { BehaviorNode, NodeStatus } from '../behavior-tree-types';
 import { Sequence, ConditionNode, ActionNode } from '../nodes';
